Migrates tests for intervals::{absolute, relative}::{bounded_interval, half_bounded_interval, interval, emptiable_interval}
