+++ src/intervals/relative_tests.rs
@@ -89,715 +89,6 @@ fn relative_start_bound_finite_relative_end_bound_finite_swap() {
 }
 
 #[test]
-fn bounded_relative_interval_unchecked_new_negative_len() {
-    let interval = BoundedRelativeInterval::unchecked_new(Duration::hours(1), Duration::hours(-5));
-
-    assert_eq!(interval.offset(), Duration::hours(1));
-    assert_eq!(interval.length(), Duration::hours(-5));
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Inclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Inclusive);
-}
-
-#[test]
-fn bounded_relative_interval_new() {
-    let interval = BoundedRelativeInterval::new(Duration::hours(1), Duration::hours(2));
-
-    assert_eq!(interval.offset(), Duration::hours(1));
-    assert_eq!(interval.length(), Duration::hours(2));
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Inclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Inclusive);
-}
-
-#[test]
-fn bounded_relative_interval_new_negative_len() {
-    let interval = BoundedRelativeInterval::new(Duration::hours(1), Duration::hours(-5));
-
-    assert_eq!(interval.offset(), Duration::hours(-4));
-    assert_eq!(interval.length(), Duration::hours(5));
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Inclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Inclusive);
-}
-
-#[test]
-#[should_panic(expected = "`TimeDelta + TimeDelta` overflowed")]
-fn bounded_relative_interval_new_negative_len_duration_underflow() {
-    let _ = BoundedRelativeInterval::new(Duration::MIN, Duration::hours(-5));
-}
-
-#[test]
-fn bounded_relative_interval_unchecked_new_with_inclusivity() {
-    let interval = BoundedRelativeInterval::unchecked_new_with_inclusivity(
-        Duration::hours(1),
-        BoundInclusivity::Inclusive,
-        Duration::zero(),
-        BoundInclusivity::Exclusive,
-    );
-
-    assert_eq!(interval.offset(), Duration::hours(1));
-    assert_eq!(interval.length(), Duration::zero());
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Inclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Exclusive);
-}
-
-#[test]
-fn bounded_relative_interval_new_with_inclusivity_zero_len() {
-    let interval = BoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(5),
-        BoundInclusivity::Inclusive,
-        Duration::zero(),
-        BoundInclusivity::Exclusive,
-    );
-
-    assert_eq!(interval.offset(), Duration::hours(5));
-    assert_eq!(interval.length(), Duration::zero());
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Inclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Inclusive);
-}
-
-#[test]
-fn bounded_relative_interval_new_with_inclusivity() {
-    let interval = BoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(1),
-        BoundInclusivity::Exclusive,
-        Duration::hours(2),
-        BoundInclusivity::Inclusive,
-    );
-
-    assert_eq!(interval.offset(), Duration::hours(1));
-    assert_eq!(interval.length(), Duration::hours(2));
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Exclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Inclusive);
-}
-
-#[test]
-fn bounded_relative_set_offset() {
-    let mut interval = BoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(1),
-        BoundInclusivity::Exclusive,
-        Duration::hours(2),
-        BoundInclusivity::Inclusive,
-    );
-
-    interval.set_offset(Duration::hours(3));
-
-    assert_eq!(interval.offset(), Duration::hours(3));
-    assert_eq!(interval.length(), Duration::hours(2));
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Exclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Inclusive);
-}
-
-#[test]
-fn bounded_relative_set_length() {
-    let mut interval = BoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(2),
-        BoundInclusivity::Exclusive,
-        Duration::hours(3),
-        BoundInclusivity::Inclusive,
-    );
-
-    interval.set_length(Duration::hours(1));
-
-    assert_eq!(interval.offset(), Duration::hours(2));
-    assert_eq!(interval.length(), Duration::hours(1));
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Exclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Inclusive);
-}
-
-#[test]
-fn bounded_relative_set_from_inclusivity() {
-    let mut interval = BoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(1),
-        BoundInclusivity::Exclusive,
-        Duration::hours(2),
-        BoundInclusivity::Inclusive,
-    );
-
-    interval.set_from_inclusivity(BoundInclusivity::Inclusive);
-
-    assert_eq!(interval.offset(), Duration::hours(1));
-    assert_eq!(interval.length(), Duration::hours(2));
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Inclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Inclusive);
-}
-
-#[test]
-fn bounded_relative_set_to_inclusivity() {
-    let mut interval = BoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(2),
-        BoundInclusivity::Exclusive,
-        Duration::hours(3),
-        BoundInclusivity::Inclusive,
-    );
-
-    interval.set_to_inclusivity(BoundInclusivity::Exclusive);
-
-    assert_eq!(interval.offset(), Duration::hours(2));
-    assert_eq!(interval.length(), Duration::hours(3));
-    assert_eq!(interval.from_inclusivity(), BoundInclusivity::Exclusive);
-    assert_eq!(interval.to_inclusivity(), BoundInclusivity::Exclusive);
-}
-
-#[test]
-fn bounded_relative_interval_from_datetime_pair() {
-    assert_eq!(
-        BoundedRelativeInterval::from((Duration::hours(2), Duration::hours(1))),
-        BoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(2),
-            BoundInclusivity::Inclusive,
-            Duration::hours(1),
-            BoundInclusivity::Inclusive,
-        ),
-    );
-}
-
-#[test]
-fn bounded_relative_interval_from_pair_of_datetime_inclusivity_pairs() {
-    assert_eq!(
-        BoundedRelativeInterval::from((
-            (Duration::hours(2), BoundInclusivity::Exclusive),
-            (Duration::hours(1), BoundInclusivity::Inclusive),
-        )),
-        BoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(2),
-            BoundInclusivity::Exclusive,
-            Duration::hours(1),
-            BoundInclusivity::Inclusive,
-        ),
-    );
-}
-
-#[test]
-fn bounded_relative_interval_from_pair_of_datetime_bool_pairs() {
-    assert_eq!(
-        BoundedRelativeInterval::from(((Duration::hours(2), false), (Duration::hours(1), true),)),
-        BoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(2),
-            BoundInclusivity::Exclusive,
-            Duration::hours(1),
-            BoundInclusivity::Inclusive,
-        ),
-    );
-}
-
-#[test]
-fn bounded_relative_interval_from_range() {
-    assert_eq!(
-        BoundedRelativeInterval::from(Duration::hours(1)..Duration::hours(2)),
-        BoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Inclusive,
-            Duration::hours(2),
-            BoundInclusivity::Exclusive,
-        ),
-    );
-}
-
-#[test]
-fn bounded_relative_interval_from_range_inclusive() {
-    assert_eq!(
-        BoundedRelativeInterval::from(Duration::hours(1)..=Duration::hours(2)),
-        BoundedRelativeInterval::new(Duration::hours(1), Duration::hours(2)),
-    );
-}
-
-#[test]
-fn bounded_relative_interval_try_from_relative_bounds_correct() {
-    assert_eq!(
-        BoundedRelativeInterval::try_from(RelativeBounds::new(
-            RelativeStartBound::Finite(RelativeFiniteBound::new_with_inclusivity(
-                Duration::hours(1),
-                BoundInclusivity::Exclusive,
-            )),
-            RelativeEndBound::Finite(RelativeFiniteBound::new_with_inclusivity(
-                Duration::hours(2),
-                BoundInclusivity::Inclusive,
-            )),
-        )),
-        Ok(BoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            Duration::hours(2),
-            BoundInclusivity::Inclusive,
-        )),
-    );
-}
-
-#[test]
-fn bounded_relative_interval_try_from_relative_bounds_wrong() {
-    assert_eq!(
-        BoundedRelativeInterval::try_from(RelativeBounds::new(
-            RelativeStartBound::InfinitePast,
-            RelativeEndBound::Finite(RelativeFiniteBound::new(Duration::hours(1))),
-        )),
-        Err(BoundedRelativeIntervalFromRelativeBoundsError::NotBoundedInterval),
-    );
-    assert_eq!(
-        BoundedRelativeInterval::try_from(RelativeBounds::new(
-            RelativeStartBound::Finite(RelativeFiniteBound::new(Duration::hours(1))),
-            RelativeEndBound::InfiniteFuture,
-        )),
-        Err(BoundedRelativeIntervalFromRelativeBoundsError::NotBoundedInterval),
-    );
-    assert_eq!(
-        BoundedRelativeInterval::try_from(RelativeBounds::new(
-            RelativeStartBound::InfinitePast,
-            RelativeEndBound::InfiniteFuture,
-        )),
-        Err(BoundedRelativeIntervalFromRelativeBoundsError::NotBoundedInterval),
-    );
-}
-
-#[test]
-fn bounded_relative_interval_try_from_relative_interval_correct() {
-    assert_eq!(
-        BoundedRelativeInterval::try_from(RelativeInterval::Bounded(BoundedRelativeInterval::new(
-            Duration::hours(1),
-            Duration::hours(2),
-        ))),
-        Ok(BoundedRelativeInterval::new(Duration::hours(1), Duration::hours(2),)),
-    );
-}
-
-#[test]
-fn bounded_relative_interval_try_from_relative_interval_wrong() {
-    assert_eq!(
-        BoundedRelativeInterval::try_from(RelativeInterval::Empty(EmptyInterval)),
-        Err(BoundedRelativeIntervalFromRelativeIntervalError::WrongVariant),
-    );
-    assert_eq!(
-        BoundedRelativeInterval::try_from(RelativeInterval::Unbounded(UnboundedInterval)),
-        Err(BoundedRelativeIntervalFromRelativeIntervalError::WrongVariant),
-    );
-    assert_eq!(
-        BoundedRelativeInterval::try_from(RelativeInterval::HalfBounded(HalfBoundedRelativeInterval::new(
-            Duration::hours(1),
-            OpeningDirection::ToFuture,
-        ))),
-        Err(BoundedRelativeIntervalFromRelativeIntervalError::WrongVariant),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_new() {
-    let interval = HalfBoundedRelativeInterval::new(Duration::hours(1), OpeningDirection::ToFuture);
-
-    assert_eq!(interval.reference_offset(), Duration::hours(1));
-    assert_eq!(interval.opening_direction(), OpeningDirection::ToFuture);
-    assert_eq!(interval.reference_inclusivity(), BoundInclusivity::Inclusive);
-}
-
-#[test]
-fn half_bounded_relative_interval_new_with_inclusivity() {
-    let interval = HalfBoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(1),
-        BoundInclusivity::Exclusive,
-        OpeningDirection::ToPast,
-    );
-
-    assert_eq!(interval.reference_offset(), Duration::hours(1));
-    assert_eq!(interval.opening_direction(), OpeningDirection::ToPast);
-    assert_eq!(interval.reference_inclusivity(), BoundInclusivity::Exclusive);
-}
-
-#[test]
-fn half_bounded_relative_interval_set_reference_time() {
-    let mut interval = HalfBoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(1),
-        BoundInclusivity::Exclusive,
-        OpeningDirection::ToFuture,
-    );
-
-    interval.set_offset(Duration::hours(2));
-
-    assert_eq!(
-        interval,
-        HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(2),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToFuture,
-        )
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_set_reference_time_inclusivity() {
-    let mut interval = HalfBoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(1),
-        BoundInclusivity::Exclusive,
-        OpeningDirection::ToFuture,
-    );
-
-    interval.set_reference_inclusivity(BoundInclusivity::Inclusive);
-
-    assert_eq!(
-        interval,
-        HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Inclusive,
-            OpeningDirection::ToFuture,
-        )
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_set_opening_direction() {
-    let mut interval = HalfBoundedRelativeInterval::new_with_inclusivity(
-        Duration::hours(1),
-        BoundInclusivity::Exclusive,
-        OpeningDirection::ToFuture,
-    );
-
-    interval.set_opening_direction(OpeningDirection::ToPast);
-
-    assert_eq!(
-        interval,
-        HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToPast,
-        )
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_from_datetime_opening_direction_pair() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::from((Duration::hours(1), OpeningDirection::ToFuture)),
-        HalfBoundedRelativeInterval::new(Duration::hours(1), OpeningDirection::ToFuture),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_from_datetime_bool_pair() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::from((Duration::hours(1), false)),
-        HalfBoundedRelativeInterval::new(Duration::hours(1), OpeningDirection::ToPast),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_from_pair_of_datetime_bound_inclusivity_pair_and_opening_direction() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::from((
-            (Duration::hours(1), BoundInclusivity::Exclusive),
-            OpeningDirection::ToPast
-        )),
-        HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToPast,
-        ),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_from_pair_of_datetime_bool_pair_and_opening_direction() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::from(((Duration::hours(1), false), OpeningDirection::ToPast)),
-        HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToPast,
-        ),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_from_pair_of_datetime_bool_pair_and_bool() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::from(((Duration::hours(1), false), false)),
-        HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToPast,
-        ),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_from_range_from() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::from(Duration::hours(1)..),
-        HalfBoundedRelativeInterval::new(Duration::hours(1), OpeningDirection::ToFuture),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_from_range_to() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::from(..Duration::hours(1)),
-        HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToPast,
-        ),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_from_range_to_inclusive() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::from(..=Duration::hours(1)),
-        HalfBoundedRelativeInterval::new(Duration::hours(1), OpeningDirection::ToPast),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_try_from_relative_bounds_correct() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::try_from(RelativeBounds::new(
-            RelativeStartBound::Finite(RelativeFiniteBound::new_with_inclusivity(
-                Duration::hours(1),
-                BoundInclusivity::Exclusive,
-            )),
-            RelativeEndBound::InfiniteFuture,
-        )),
-        Ok(HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToFuture,
-        )),
-    );
-    assert_eq!(
-        HalfBoundedRelativeInterval::try_from(RelativeBounds::new(
-            RelativeStartBound::InfinitePast,
-            RelativeEndBound::Finite(RelativeFiniteBound::new_with_inclusivity(
-                Duration::hours(1),
-                BoundInclusivity::Exclusive,
-            )),
-        )),
-        Ok(HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToPast,
-        )),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_try_from_relative_bounds_wrong() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::try_from(RelativeBounds::new(
-            RelativeStartBound::InfinitePast,
-            RelativeEndBound::InfiniteFuture,
-        )),
-        Err(HalfBoundedRelativeIntervalFromRelativeBoundsError::NotHalfBoundedInterval),
-    );
-    assert_eq!(
-        HalfBoundedRelativeInterval::try_from(RelativeBounds::new(
-            RelativeStartBound::Finite(RelativeFiniteBound::new(Duration::hours(1))),
-            RelativeEndBound::Finite(RelativeFiniteBound::new(Duration::hours(2))),
-        )),
-        Err(HalfBoundedRelativeIntervalFromRelativeBoundsError::NotHalfBoundedInterval),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_try_from_relative_interval_correct() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::try_from(RelativeInterval::HalfBounded(
-            HalfBoundedRelativeInterval::new_with_inclusivity(
-                Duration::hours(1),
-                BoundInclusivity::Exclusive,
-                OpeningDirection::ToPast,
-            )
-        )),
-        Ok(HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToPast,
-        )),
-    );
-}
-
-#[test]
-fn half_bounded_relative_interval_try_from_relative_interval_wrong() {
-    assert_eq!(
-        HalfBoundedRelativeInterval::try_from(RelativeInterval::Empty(EmptyInterval)),
-        Err(HalfBoundedRelativeIntervalFromRelativeIntervalError::WrongVariant),
-    );
-    assert_eq!(
-        HalfBoundedRelativeInterval::try_from(RelativeInterval::Unbounded(UnboundedInterval)),
-        Err(HalfBoundedRelativeIntervalFromRelativeIntervalError::WrongVariant),
-    );
-    assert_eq!(
-        HalfBoundedRelativeInterval::try_from(RelativeInterval::Bounded(BoundedRelativeInterval::new(
-            Duration::hours(1),
-            Duration::hours(2),
-        ))),
-        Err(HalfBoundedRelativeIntervalFromRelativeIntervalError::WrongVariant),
-    );
-}
-
-#[test]
-fn relative_interval_from_relative_bounds() {
-    assert_eq!(
-        RelativeInterval::from(RelativeBounds::new(
-            RelativeStartBound::Finite(RelativeFiniteBound::new_with_inclusivity(
-                Duration::hours(1),
-                BoundInclusivity::Exclusive,
-            )),
-            RelativeEndBound::InfiniteFuture,
-        )),
-        RelativeInterval::HalfBounded(HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToFuture,
-        )),
-    );
-}
-
-#[test]
-fn relative_interval_from_emptiable_relative_bounds() {
-    assert_eq!(
-        RelativeInterval::from(EmptiableRelativeBounds::Bound(RelativeBounds::new(
-            RelativeStartBound::Finite(RelativeFiniteBound::new_with_inclusivity(
-                Duration::hours(1),
-                BoundInclusivity::Exclusive,
-            )),
-            RelativeEndBound::InfiniteFuture,
-        ))),
-        RelativeInterval::HalfBounded(HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToFuture,
-        )),
-    );
-}
-
-#[test]
-fn relative_interval_from_opt_datetime_pair_unbounded() {
-    assert_eq!(
-        <RelativeInterval as From<(Option<Duration>, Option<Duration>)>>::from((None, None)),
-        RelativeInterval::Unbounded(UnboundedInterval),
-    );
-}
-
-#[test]
-fn relative_interval_from_opt_datetime_pair_half_bounded() {
-    assert_eq!(
-        RelativeInterval::from((None, Some(Duration::hours(1)))),
-        RelativeInterval::HalfBounded(HalfBoundedRelativeInterval::new(
-            Duration::hours(1),
-            OpeningDirection::ToPast,
-        )),
-    );
-}
-
-#[test]
-fn relative_interval_from_opt_datetime_bound_inclusivity_pairs() {
-    assert_eq!(
-        RelativeInterval::from((
-            Some((Duration::hours(1), BoundInclusivity::Exclusive)),
-            Some((Duration::hours(2), BoundInclusivity::Exclusive)),
-        )),
-        RelativeInterval::Bounded(BoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            Duration::hours(2),
-            BoundInclusivity::Exclusive,
-        )),
-    );
-}
-
-#[test]
-fn relative_interval_from_opt_datetime_bool_pairs() {
-    assert_eq!(
-        RelativeInterval::from((Some((Duration::hours(1), true)), Some((Duration::hours(2), false)),)),
-        RelativeInterval::Bounded(BoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Inclusive,
-            Duration::hours(2),
-            BoundInclusivity::Exclusive,
-        )),
-    );
-}
-
-#[test]
-fn relative_interval_from_bool_and_two_opt_datetime_empty() {
-    assert_eq!(
-        <RelativeInterval as From<(bool, Option<Duration>, Option<Duration>)>>::from((true, None, None,)),
-        RelativeInterval::Empty(EmptyInterval),
-    );
-}
-
-#[test]
-fn relative_interval_from_bool_and_two_opt_datetime() {
-    assert_eq!(
-        RelativeInterval::from((false, Some(Duration::hours(1)), Some(Duration::hours(2)),)),
-        RelativeInterval::Bounded(BoundedRelativeInterval::new(Duration::hours(1), Duration::hours(2))),
-    );
-}
-
-#[test]
-fn relative_interval_from_bool_and_two_opt_datetime_bound_inclusivity_empty() {
-    assert_eq!(
-        <RelativeInterval as From<(
-            bool,
-            Option<(Duration, BoundInclusivity)>,
-            Option<(Duration, BoundInclusivity)>
-        )>>::from((true, None, None,)),
-        RelativeInterval::Empty(EmptyInterval),
-    );
-}
-
-#[test]
-fn relative_interval_from_bool_and_two_opt_datetime_bound_inclusivity() {
-    assert_eq!(
-        RelativeInterval::from((
-            false,
-            Some((Duration::hours(1), BoundInclusivity::Exclusive)),
-            Some((Duration::hours(2), BoundInclusivity::Exclusive)),
-        )),
-        RelativeInterval::Bounded(BoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            Duration::hours(2),
-            BoundInclusivity::Exclusive,
-        )),
-    );
-}
-
-#[test]
-fn relative_interval_from_bool_and_two_opt_datetime_bool_empty() {
-    assert_eq!(
-        <RelativeInterval as From<(bool, Option<(Duration, bool)>, Option<(Duration, bool)>)>>::from((
-            true, None, None,
-        )),
-        RelativeInterval::Empty(EmptyInterval),
-    );
-}
-
-#[test]
-fn relative_interval_from_bool_and_two_opt_datetime_bool() {
-    assert_eq!(
-        RelativeInterval::from((
-            false,
-            Some((Duration::hours(1), false)),
-            Some((Duration::hours(2), false)),
-        )),
-        RelativeInterval::Bounded(BoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            Duration::hours(2),
-            BoundInclusivity::Exclusive,
-        )),
-    );
-}
-
-#[test]
-fn relative_interval_from_bound_pair() {
-    assert_eq!(
-        RelativeInterval::from((Bound::Unbounded, Bound::Excluded(Duration::hours(1)))),
-        RelativeInterval::HalfBounded(HalfBoundedRelativeInterval::new_with_inclusivity(
-            Duration::hours(1),
-            BoundInclusivity::Exclusive,
-            OpeningDirection::ToPast,
-        )),
-    );
-}
-
-#[test]
 fn check_relative_bounds_for_interval_creation_inf_past_inf_future() {
     assert_eq!(
         check_relative_bounds_for_interval_creation(
